Set up pre-test deletion and env-specific post test deletion for ProductSpecification mutations

+++ src/ecomm-api-svc/cypress/integration/mutations/productSpecification/deleteProductSpecification.spec.ts
@@ -29,6 +29,12 @@ describe('Mutation: deleteProductSpecification', () => {
         queryInformation.itemName = providedName ? providedName : "";
     };
 
+	var deleteItemsAfter = undefined as boolean | undefined;
+	before(() => {
+		deleteItemsAfter = Cypress.env("deleteItemsAfter");
+		cy.deleteCypressItems(queryName, mutationName);
+	});
+
     beforeEach(() => {
         const name = `Cypress test: ${mutationName}'s deletee`;
         const input = `{name: "${name}", options: [{name: "PA deletee option"}]}`;
@@ -38,6 +44,9 @@ describe('Mutation: deleteProductSpecification', () => {
     });
 
     afterEach(() => {
+		if (!deleteItemsAfter) {
+			return;
+		}
         // Delete any supplemental items we created
         cy.deleteSupplementalItems(extraIds).then(() => {
             extraIds = [];
